Rename `_state_managment` to `_state_management` (E-07)
The private module `braintrace/_state_managment.py` has been misspelled
since it was introduced ("managment" is missing an `e`). A wrong file
name is load-bearing in a way a wrong comment is not: it is what every
import line has to repeat, so the typo propagated into three source
modules, its own test, and the mypy per-module configuration list.

Renamed via `git mv`, together with its co-located test
(`AGENTS.md` rule 9). Module contents are unchanged; the four exported
helpers keep their names and behaviour.

No deprecation shim and no DeprecationWarning. The module is private,
none of its helpers is re-exported from `braintrace`, and a shim has no
natural removal trigger — it would keep the misspelling importable, and
therefore greppable and copy-pasteable, indefinitely. The break is
narrow but real, so it is announced in `changelog.md` rather than
absorbed silently.

Reference sites updated: `_algorithm/vjp_base.py`,
`_algorithm/vjp_graph_executor.py`, `_compiler/module_info.py`, the
renamed test, and the `disallow_untyped_defs` module list in
`pyproject.toml`. Mentions of the old spelling that are records of a
past state — the 0.1.0 release notes and the E-07 audit entry — are
left in their original wording; only pointers at live code were
rewritten.

Spec: `docs/specs/2026-08-07-e07-state-management-rename.md`
