+++ docs/specs/2026-08-07-e07-state-management-rename.md
@@ -1,0 +1,132 @@
+# E-07 — rename `_state_managment.py` to `_state_management.py`
+
+Status: implemented
+Scope: `braintrace/_state_managment.py` → `braintrace/_state_management.py`
+Backlog entry: E-07 in `2026-08-07-deferred-engineering-backlog.md`
+Issue: [#162](https://github.com/chaobrain/braintrace/issues/162)
+
+## The defect
+
+The module `braintrace/_state_managment.py` has been misspelled since it was
+introduced ("managment" is missing the second `e`). Nothing about its contents
+is wrong — the four helpers it exports (`assign_dict_state_values`,
+`assign_state_values_v2`, `sequence_split_state_values`,
+`split_dict_states_v2`) keep their names and their behaviour. Only the file name
+is wrong, and a wrong file name is load-bearing in a way a wrong comment is not:
+it is what every `from braintrace._state_managment import ...` line has to
+repeat, so the typo propagates into every importing module and into the mypy
+per-module configuration list.
+
+## The change
+
+A clean rename, recorded through `git mv` so history follows the file:
+
+- `braintrace/_state_managment.py` → `braintrace/_state_management.py`
+- `braintrace/_state_managment_test.py` → `braintrace/_state_management_test.py`
+  (`AGENTS.md` rule 9 — the test stays a sibling of the module under test, so it
+  has to move with it)
+
+The module body is untouched apart from the import line inside its own test.
+
+## No deprecation shim — the decision and why
+
+The backlog entry left the choice open between a re-exporting shim and accepting
+the break. The decision taken here is **accept the break, ship no shim**:
+
+- The module is private. The leading underscore is the package's own statement
+  that the path is not part of the supported surface, and none of the four
+  helpers is re-exported from `braintrace/__init__.py` or from any `nn`
+  namespace. Anyone importing `braintrace._state_managment` was already reaching
+  past the public API and has accepted that the path can move.
+- A shim would be permanent in practice. A back-compat module that merely
+  re-exports has no natural removal trigger — it costs one more file, one more
+  import edge and one more entry in every module list, forever, to preserve a
+  path nobody was promised.
+- A shim would preserve the typo as a *supported* spelling, which is the exact
+  thing the issue exists to remove. Keeping `_state_managment` importable means
+  the misspelling stays greppable, autocompletable and copy-pasteable.
+- No `DeprecationWarning` either, for the same reason: warning about a private
+  path implies the path had a deprecation contract to begin with.
+
+The break is therefore real but narrow, and it is announced in `changelog.md`
+rather than absorbed silently.
+
+## Reference sites updated
+
+Live code and configuration — every one of these had to change or the package
+would not import:
+
+| File | What referenced the old name |
+| --- | --- |
+| `braintrace/_algorithm/vjp_base.py` | `from braintrace._state_managment import assign_state_values_v2` |
+| `braintrace/_algorithm/vjp_graph_executor.py` | `from braintrace._state_managment import (assign_dict_state_values, split_dict_states_v2)` |
+| `braintrace/_compiler/module_info.py` | `from braintrace._state_managment import sequence_split_state_values` |
+| `braintrace/_state_management_test.py` | its own import of the module under test |
+| `pyproject.toml` | the `[[tool.mypy.overrides]]` `disallow_untyped_defs = true` module list |
+
+`pyproject.toml` was checked for other module-path patterns: the mypy `exclude`
+list and the `[tool.coverage.run]` `omit` list are glob patterns
+(`*_test.py`, `*/model4test.py`, …) that do not name this module, and
+`[tool.setuptools]` uses package-level discovery rather than a module list, so
+nothing else there needed touching. `setup.py`, `.github/workflows/`,
+`examples/`, `docs/` (`.md`, `.rst`, `.ipynb`, all `automodule::` directives)
+and `conftest.py` contain no reference to the module under either spelling —
+the docs mention "state management" only as English prose. There is no
+`dev/` directory in this repository.
+
+## Historical records deliberately left alone
+
+After this change,
+
+```console
+$ grep -rn "_state_managment" . --exclude-dir=.git --exclude=changelog.md --exclude-dir=docs/specs
+$
+```
+
+returns nothing: no source file, no test, no configuration file and no
+user-facing document mentions the old spelling. The string survives only in
+Markdown that is *about* the old spelling, which cannot avoid naming it — this
+spec, the new `changelog.md` entry announcing the rename, and two records that
+are deliberately left alone:
+
+- `changelog.md` line ~1239 — the 0.1.0 release notes list the files touched by
+  the `ETraceState` → `HiddenState` refactor. That file *was* called
+  `_state_managment.py` in 0.1.0. Rewriting the entry would make a past release
+  note describe a file that did not exist at that release.
+- `docs/specs/2026-08-07-deferred-engineering-backlog.md` — the E-07 entry, whose
+  entire subject is the misspelling. Renaming the string inside it would reduce
+  the heading to "`_state_management.py` is misspelled", which is nonsense. The
+  entry is instead annotated as resolved, with a pointer to this spec and to the
+  release that shipped the rename.
+
+The rule applied: a mention of the old name is rewritten when it is a *pointer at
+live code* (an import, a config entry, an autodoc target) and preserved when it
+is a *record of what was true at a past point in time*. Every pointer at live
+code has been rewritten; only records remain.
+
+## Edge cases considered
+
+- **Stale bytecode.** A `__pycache__/_state_managment.*.pyc` left in a working
+  tree does not make the old path importable — CPython requires the matching
+  `.py` source for a cached module, so the import fails cleanly rather than
+  resolving to a stale cache.
+- **Case-insensitive filesystems.** Not a factor: the two names differ by an
+  inserted character, not by case, so this is an ordinary rename on Windows and
+  macOS as well as Linux.
+- **Editable installs.** The package is discovered by directory, not by an
+  explicit module list, so an existing editable install picks up the new name
+  with no reinstall.
+- **Wheel contents.** One file name changes; the set of importable public
+  symbols is identical.
+
+## Tests
+
+No new test is written for the rename itself. A rename has no behaviour to
+assert beyond "the package imports and everything that used it still works",
+which the existing suite already asserts far more thoroughly than a bespoke test
+could: `braintrace/_state_management_test.py` exercises all four helpers
+through the new path, and `braintrace/__init___test.py` plus the algorithm and
+compiler suites fail at import time if any reference were missed. A test
+asserting that `braintrace._state_managment` is *not* importable was considered
+and rejected — it would pin the absence of a shim as a contract, which is more
+commitment than a private path deserves.
